# planning/api-documentations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "echomemory backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/music/admin/import": {
      "post": {
        "tags": [
          "music"
        ],
        "summary": "Import Music",
        "description": "管理员导入音乐。\n\n所有文件均通过上传方式提供，服务端自动上传到 OSS 并生成 URL。\n禁止直接填写任何 URL 字符串。",
        "operationId": "import_music_api_v1_music_admin_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_music_api_v1_music_admin_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/music/admin/{music_id}": {
      "patch": {
        "tags": [
          "music"
        ],
        "summary": "Admin Update Music",
        "description": "管理员修改音乐信息（不含文件替换）。",
        "operationId": "admin_update_music_api_v1_music_admin__music_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/admin/{music_id}/publish": {
      "post": {
        "tags": [
          "music"
        ],
        "summary": "Admin Publish Music",
        "description": "管理员上架音乐。",
        "operationId": "admin_publish_music_api_v1_music_admin__music_id__publish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/admin/{music_id}/unpublish": {
      "post": {
        "tags": [
          "music"
        ],
        "summary": "Admin Unpublish Music",
        "description": "管理员下架音乐。",
        "operationId": "admin_unpublish_music_api_v1_music_admin__music_id__unpublish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/search": {
      "get": {
        "tags": [
          "music"
        ],
        "summary": "Search Musics",
        "description": "按标题模糊搜索已上架音乐。",
        "operationId": "search_musics_api_v1_music_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按标题模糊搜索",
              "title": "Q"
            },
            "description": "按标题模糊搜索"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MusicListOut"
                  },
                  "title": "Response Search Musics Api V1 Music Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/": {
      "get": {
        "tags": [
          "music"
        ],
        "summary": "List Musics",
        "description": "分页列出已上架音乐，支持筛选。",
        "operationId": "list_musics_api_v1_music__get",
        "parameters": [
          {
            "name": "style_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Style Id"
            }
          },
          {
            "name": "language_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language Id"
            }
          },
          {
            "name": "is_vip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Vip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MusicListOut"
                  },
                  "title": "Response List Musics Api V1 Music  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/music/{music_id}": {
      "get": {
        "tags": [
          "music"
        ],
        "summary": "Get Music",
        "description": "获取已上架音乐的详情。",
        "operationId": "get_music_api_v1_music__music_id__get",
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/playlists/{playlist_id}/musics/{music_id}": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Add Music To Playlist",
        "description": "添加一首已上架音乐到歌单。仅允许操作自己的歌单。",
        "operationId": "add_music_to_playlist_api_v1_playlists__playlist_id__musics__music_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Remove Music From Playlist",
        "description": "从歌单移除一首音乐。仅允许操作自己的歌单。",
        "operationId": "remove_music_from_playlist_api_v1_playlists__playlist_id__musics__music_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/albums/admin/{album_id}/musics/{music_id}": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Add Music To Album",
        "description": "将一首已上架音乐加入专辑。",
        "operationId": "add_music_to_album_api_v1_albums_admin__album_id__musics__music_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "albums"
        ],
        "summary": "Remove Music From Album",
        "description": "从专辑移除一首音乐。",
        "operationId": "remove_music_from_album_api_v1_albums_admin__album_id__musics__music_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/musics/{music_id}": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Collect Music",
        "description": "收藏音乐。已收藏则静默返回。",
        "operationId": "collect_music_api_v1_collections_musics__music_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicCollectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Uncollect Music",
        "description": "取消收藏音乐。未收藏则静默成功。",
        "operationId": "uncollect_music_api_v1_collections_musics__music_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/musics": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List Music Collections",
        "description": "查询我的收藏音乐列表（按收藏时间倒序）。",
        "operationId": "list_music_collections_api_v1_collections_musics_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MusicCollectionOut"
                  },
                  "title": "Response List Music Collections Api V1 Collections Musics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_music_api_v1_music_admin_import_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title"
          },
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File"
          },
          "cover_icon": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cover Icon"
          },
          "is_vip": {
            "type": "boolean",
            "title": "Is Vip",
            "default": false
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "style_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Id"
          },
          "language_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Id"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date",
            "description": "格式: YYYY-MM-DD"
          },
          "author_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Author Ids",
            "default": []
          },
          "instrument_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Instrument Ids",
            "default": []
          },
          "emotion_tag_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Emotion Tag Ids",
            "default": []
          },
          "interest_tag_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Interest Tag Ids",
            "default": []
          },
          "cover_home": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Home"
          },
          "cover_play": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Play"
          },
          "lyrics_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lyrics File"
          }
        },
        "type": "object",
        "required": [
          "title",
          "audio_file",
          "cover_icon"
        ],
        "title": "Body_import_music_api_v1_music_admin_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}